feat: Add SidebarRequestInternship component for internship request sidebar
The code changes include adding a new component called SidebarRequestInternship. This component is responsible for rendering the sidebar section on the RequestInternship page. It displays information about the internship request process and includes a "Kirim" button for submitting the request. This component enhances the user experience by providing clear instructions and a convenient way to submit internship requests.