feat: Update SidebarRequestInternship component
The code changes involve updating the SidebarRequestInternship component in the index.jsx file. The "Kirim" button has been modified to include a "type" attribute with the value "submit" to enable form submission. This enhancement improves the functionality of the component by allowing the form to be submitted when the button is clicked.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature and providing a clear and concise description of the changes made.