Refactor UserContext.jsx to add setUserLoggedInData function
Add the setUserLoggedInData function to the UserContext.jsx file. This function is used to set the user's logged-in data, providing a way to update the user's data in the context. This change improves the functionality of the UserContext and enhances the user experience.

Note: This commit message describes the purpose of the code changes and follows the established convention of providing a clear and concise description of the changes made.