Refactor FormRequestInternship component
Refactor the FormRequestInternship component to improve code readability and maintainability. This change includes adding a select dropdown for selecting a "Dosen Wali" and an input field for entering a "Nomor Whatsapp". These enhancements provide additional functionality and improve the user experience.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate code refactoring and providing a clear and concise description of the changes made.